Create ConsoleTestCase to contain console-only helpers
