+++ test/console/display_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/console_test_case'
 
 module DEBUGGER__
-  class DisplayTest < TestCase
+  class DisplayTest < ConsoleTestCase
     def program
       <<~RUBY
      1| a = 1
