+++ test/console/list_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/console_test_case'
 
 module DEBUGGER__
-  class ListTest < TestCase
+  class ListTest < ConsoleTestCase
     def program
       <<~RUBY
       1| p 1
